requirements_Jetson_mqtt.txt
# certifi==2024.2.2
# charset-normalizer==3.3.2
# contourpy==1.2.0
# cycler==0.12.1
# defusedxml==0.7.1
defusedxml
# filelock==3.13.1
# fonttools==4.49.0
# fsspec==2024.2.0
# idna==3.6
# Jinja2==3.1.3
# kiwisolver==1.4.5
# MarkupSafe==2.1.5
# matplotlib==3.8.3
# mpmath==1.3.0
# networkx==3.2.1
# numpy==1.26.4
numpy<2
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.3.101
# nvidia-nvtx-cu12==12.1.105
# opencv-python==4.6.0.66
# opencv-python-headless==4.9.0.80
# opencv-python-headless
# packaging==23.2
# pandas==2.2.0
paho-mqtt
# pillow==10.2.0
# psutil==5.9.8
psutil
# py-cpuinfo==9.0.0
# pyparsing==3.1.1
# python-dateutil==2.8.2
# pytz==2024.1
# PyYAML==6.0.1
# requests==2.31.0
# scipy==1.12.0
# seaborn==0.13.2
# six==1.16.0
# sympy==1.12
# thop==0.1.1.post2209072238
# tqdm==4.66.2
# triton==2.2.0
# typing_extensions==4.9.0
typing_extensions
# tzdata==2024.1
# ultralytics==8.1.17
ultralytics
# urllib3==2.2.1
